# scratch/aranet-openapi.json
{
  "openapi": "3.0.3",
  "components": {
    "schemas": {
      "common.BaseResponse": {
        "description": "Base API response",
        "properties": {
          "error": {
            "description": "Errors for error response (500 InternalServerError) or warnings if successful response (200 OK)",
            "items": {
              "$ref": "#/components/schemas/common.ErrorDesc"
            },
            "type": "array"
          },
          "links": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/common.LinkDesc"
              },
              "type": "array"
            },
            "description": "Links to related resources",
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "common.ErrorDesc": {
        "description": "Error",
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/common.ErrorDetails"
            },
            "type": "array"
          },
          "id": {
            "description": "Unique incident identifier",
            "type": "string"
          },
          "message": {
            "description": "Error text",
            "type": "string"
          }
        },
        "type": "object"
      },
      "common.ErrorDetails": {
        "description": "Error details",
        "properties": {
          "action": {
            "description": "Action to resolve incident",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "description": "Incident object type",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "common.LinkDesc": {
        "description": "Link to related resource",
        "properties": {
          "href": {
            "description": "Link to related resource",
            "type": "string"
          },
          "name": {
            "description": "Link name",
            "type": "string"
          },
          "rel": {
            "description": "Link relation. Use to find link by relation value",
            "type": "string"
          }
        },
        "type": "object"
      },
      "sensor.AlarmInfo": {
        "description": "Alarm information",
        "properties": {
          "alarmed": {
            "description": "Entered alarms state",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Alarm id",
            "type": "string"
          },
          "metric": {
            "description": "Metric id",
            "type": "string"
          },
          "note": {
            "description": "Alarm note",
            "type": "string"
          },
          "resolved": {
            "description": "Alarm resolved",
            "format": "date-time",
            "type": "string"
          },
          "rule": {
            "description": "Alarm rule id",
            "type": "string"
          },
          "sensor": {
            "description": "Sensor id",
            "type": "string"
          },
          "severity": {
            "description": "Alarm severity",
            "type": "integer"
          },
          "threshold": {
            "description": "Threshold direction",
            "type": "string"
          },
          "unit": {
            "description": "Unit id",
            "type": "string"
          },
          "value": {
            "description": "Threshold value",
            "type": "number"
          },
          "worst": {
            "description": "Worst value during alarm period",
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "sensor.AlarmsListResponse": {
        "description": "Alarms response with collection of entities (actual or historical)",
        "properties": {
          "alarms": {
            "description": "Alarm data",
            "items": {
              "$ref": "#/components/schemas/sensor.AlarmInfo"
            },
            "type": "array"
          },
          "error": {
            "description": "Errors for error response (500 InternalServerError) or warnings if successful response (200 OK)",
            "items": {
              "$ref": "#/components/schemas/common.ErrorDesc"
            },
            "type": "array"
          },
          "links": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/common.LinkDesc"
              },
              "type": "array"
            },
            "description": "Links to related resources",
            "type": "object"
          },
          "self": {
            "description": "Self link",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
...
    "securitySchemes": {
      "ApiKey": {
        "in": "header",
        "name": "ApiKey",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Aranet Cloud API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/alarms/history": {
      "get": {
        "operationId": "Alarms.Api.History",
        "parameters": [
          {
            "description": "Time period from",
            "in": "query",
            "name": "from",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Time period to",
            "in": "query",
            "name": "to",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sensor.AlarmsListResponse"
                }
              }
            },
            "description": "Alarms collection response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common.BaseResponse"
                }
              }
            },
            "description": "Error message"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common.BaseResponse"
                }
              }
            },
            "description": "Error message"
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "Historical alarms",
        "tags": [
          "alarms"
        ]
      }
    },
...
  },
...
}